To avoid removing messages that lack text content by their very design (e.g., stickers and images), I decided to only check if chat bubble is empty and remove accordingly on the condition that the bubble contained TeX math delimiters before an attempt was made to parse its content. In order to write the corresponding conditional statement in parseParts in parse.js and avoid duplicating operations, I invoked getTexBounds exactly once for each addition of a bubble to the message grid.
Replacing empty bubbles with text-only empty bubble error messages was making scrolling janky. So I decided to leave the empty bubbles in place and simply restyle them with backgrounds matching the color of the surrounding message grid. But to accomplish that I had to rewrite the chatBubble selector in selector.js to match certain ancestors of the elements matched by the original selector. After all this, certain bubbles insisted on retaining blue backgrounds. To get the bubbles to cooperate background-wise, I replaced those nodes with clones.

I fixed an intermittent bug where the CSS files listed in web_accessible_resources could not be accessed (even in Chrome), because chrome.runtime was undefined.

I gave users the option to copy the LaTeX code that produces a given expression by right-clicking that expression and selecting "Copy LaTeX" from the menu.