Sometimes the chat box container is (apparently) not available right after you switch from Messenger view to chat box view, so wait for it to become available.
We don't want to compare the width of the formula to that of its wrapper span (which is its parent). We want to compare the width of the formula to that of the parent of the wrapper span (its grandparent). That's how we can tell whether it initially fits in the chat bubble or not.